requirements.txt
aiosmtplib==2.0.2
anyio==3.7.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.99.0
fastapi-mail==1.3.1
fastapi-sso==0.6.4
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.3
multipart==0.2.4
nodeenv==1.8.0
oauthlib==3.2.2
passlib==1.7.4
prisma==0.9.1
pyasn1==0.5.0
pydantic==1.10.10
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
starlette-admin==0.10.0
telebot==0.0.5
tomlkit==0.11.8
typing_extensions==4.7.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
